monitor/requirements.txt
httpx
httpx[http2]
motor
dnspython
python-dotenv
